# Reduce allocations in sortKeysMapEncoder