Use one buffer for all encoded keys and values.

This significantly reduces allocations in my test of an array of 18 objects and each object has six fields.

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkArrayJSONMarshal/jsoniter-8     67619         62885         -7.00%

benchmark                                old allocs     new allocs     delta
BenchmarkArrayJSONMarshal/jsoniter-8     500            393            -21.40%

benchmark                                old bytes     new bytes     delta
BenchmarkArrayJSONMarshal/jsoniter-8     30917         20421         -33.95%
```

It also brought allocations below the stdlib implementation for this test:

```
BenchmarkArrayJSONMarshal/stdlib-8         	   13418	     89073 ns/op	   26984 B/op	     466 allocs/op
```